requirements.txt
debugpy>=1.8.0,<2.0.0
fastapi>=0.112.0,<1.0
python-dotenv==1.0.1
python-jose>=3.3.0,<4.0
sentence-transformers>=3.0.1,<4.0
uvicorn>=0.30.5,<=0.31.0